group project features-related endpoints within their own controller and OpenAPI category
