+++ api/apps/api/src/modules/projects/projects.controller.ts
@@ -13,7 +13,6 @@ import {
   ParseUUIDPipe,
   Patch,
   Post,
-  Query,
   Req,
   Res,
   UploadedFile,
@@ -24,16 +23,13 @@ import {
 import { projectResource, ProjectResultSingular } from './project.api.entity';
 import {
   ApiBearerAuth,
-  ApiBody,
   ApiConsumes,
   ApiCreatedResponse,
-  ApiForbiddenResponse,
   ApiOkResponse,
   ApiOperation,
   ApiParam,
   ApiProduces,
   ApiTags,
-  ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
 import { apiGlobalPrefixes } from '@marxan-api/api.config';
 import { JwtAuthGuard } from '@marxan-api/guards/jwt-auth.guard';
@@ -43,18 +39,9 @@ import {
   uploadOptions,
 } from '@marxan-api/utils/file-uploads.utils';
 
-import { JSONAPIQueryParams } from '@marxan-api/decorators/json-api-parameters.decorator';
 import { UpdateProjectDTO } from './dto/update.project.dto';
 import { CreateProjectDTO } from './dto/create.project.dto';
 import { RequestWithAuthenticatedUser } from '@marxan-api/app.controller';
-import {
-  FetchSpecification,
-  ProcessFetchSpecification,
-} from 'nestjs-base-service';
-import {
-  GeoFeature,
-  GeoFeatureResult,
-} from '@marxan-api/modules/geo-features/geo-feature.api.entity';
 import { ApiConsumesShapefile } from '../../decorators/shapefile.decorator';
 import {
   notAllowed,
@@ -62,16 +49,11 @@ import {
   ProjectsService,
   validationFailed,
 } from './projects.service';
-import { GeoFeatureSerializer } from './dto/geo-feature.serializer';
 import { ProjectSerializer } from './dto/project.serializer';
 import { ProjectJobsStatusDto } from './dto/project-jobs-status.dto';
 import { JobStatusSerializer } from './dto/job-status.serializer';
 import { PlanningAreaResponseDto } from './dto/planning-area-response.dto';
 import { isLeft } from 'fp-ts/Either';
-import { UploadShapefileDTO } from './dto/upload-shapefile.dto';
-import { GeoFeaturesService } from '@marxan-api/modules/geo-features';
-import { ShapefileService } from '@marxan/shapefile-converter';
-import { isFeatureCollection } from '@marxan/utils';
 import { asyncJobTag } from '@marxan-api/dto/async-job-tag';
 import { inlineJobTag } from '@marxan-api/dto/inline-job-tag';
 import { UpdateProjectBlmRangeDTO } from '@marxan-api/modules/projects/dto/update-project-blm-range.dto';
@@ -121,18 +103,7 @@ import {
   unfinishedExport,
 } from '../clone/export/application/get-archive.query';
 import { deleteProjectFailed } from './delete-project/delete-project.command';
-import {
-  featureIsLinkedToOneOrMoreScenarios,
-  featureNotDeletable,
-  featureNotFound,
-} from '@marxan-api/modules/geo-features/geo-features.service';
-import { ApiConsumesCsv } from '@marxan-api/decorators/csv.decorator';
-import { UpdateGeoFeatureTagDTO } from '@marxan-api/modules/geo-feature-tags/dto/update-geo-feature-tag.dto';
-import { GeoFeatureTagsService } from '@marxan-api/modules/geo-feature-tags/geo-feature-tags.service';
-import { GetProjectTagsResponseDto } from '@marxan-api/modules/projects/dto/get-project-tags-response.dto';
-import { UpdateProjectTagDTO } from '@marxan-api/modules/projects/dto/update-project-tag.dto';
 import { outputProjectSummaryResource } from './output-project-summaries/output-project-summary.api.entity';
-import { isNil } from 'lodash';
 import { UploadShapefileDto } from '@marxan-api/modules/scenarios/dto/upload.shapefile.dto';
 import {
   AsyncJobDto,
@@ -147,49 +118,10 @@ import { scenarioResource } from '@marxan-api/modules/scenarios/scenario.api.ent
 export class ProjectsController {
   constructor(
     private readonly projectsService: ProjectsService,
-    private readonly geoFeatureSerializer: GeoFeatureSerializer,
-    private readonly geoFeatureService: GeoFeaturesService,
-    private readonly geoFeatureTagsService: GeoFeatureTagsService,
     private readonly projectSerializer: ProjectSerializer,
     private readonly jobsStatusSerializer: JobStatusSerializer,
-    private readonly shapefileService: ShapefileService,
     private readonly proxyService: ProxyService,
   ) {}
-
-  @IsMissingAclImplementation()
-  @ApiOperation({
-    description: 'Find all geo features',
-  })
-  @ApiOkResponse({
-    type: GeoFeatureResult,
-  })
-  @ApiUnauthorizedResponse()
-  @ApiForbiddenResponse()
-  @JSONAPIQueryParams()
-  @Get(':projectId/features')
-  async findAllGeoFeaturesForProject(
-    @ProcessFetchSpecification() fetchSpecification: FetchSpecification,
-    @Param('projectId', ParseUUIDPipe) projectId: string,
-    @Req() req: RequestWithAuthenticatedUser,
-    @Query('q') featureClassAndAliasFilter?: string,
-  ): Promise<GeoFeatureResult> {
-    const result = await this.projectsService.findAllGeoFeatures(
-      fetchSpecification,
-      {
-        authenticatedUser: req.user,
-        params: {
-          projectId: projectId,
-          featureClassAndAliasFilter: featureClassAndAliasFilter,
-        },
-      },
-    );
-    if (isLeft(result)) {
-      throw new NotFoundException();
-    }
-    const { data, metadata } = result.right;
-
-    return this.geoFeatureSerializer.serialize(data, metadata);
-  }
 
   @ImplementsAcl()
   @ApiOperation({ description: 'Create project' })
@@ -490,308 +422,6 @@ export class ProjectsController {
     req.query = result.right.query;
 
     return await this.proxyService.proxyTileRequest(req, response);
-  }
-
-  @IsMissingAclImplementation()
-  @ApiConsumesShapefile({ withGeoJsonResponse: false })
-  @ApiOperation({
-    description: `Upload shapefiles of species or bioregional features.`,
-  })
-  @ApiOkResponse({ type: GeoFeature })
-  @ApiTags(inlineJobTag)
-  @ApiBody({
-    description: 'Shapefile to upload',
-    type: UploadShapefileDTO,
-  })
-  @Post(`:id/features/shapefile`)
-  @GeometryFileInterceptor(GeometryKind.ComplexWithProperties)
-  async uploadFeatures(
-    @Param('id') projectId: string,
-    @UploadedFile() shapefile: Express.Multer.File,
-    @Body() body: UploadShapefileDTO,
-  ): Promise<GeoFeature> {
-    await ensureShapefileHasRequiredFiles(shapefile);
-
-    const { data } = await this.shapefileService.transformToGeoJson(shapefile, {
-      allowOverlaps: true,
-    });
-
-    if (!isFeatureCollection(data)) {
-      throw new BadRequestException(`Only FeatureCollection is supported.`);
-    }
-
-    const newFeatureOrError = await this.geoFeatureService.createFeaturesForShapefile(
-      projectId,
-      body,
-      data.features,
-    );
-
-    if (isLeft(newFeatureOrError)) {
-      // @debt Use mapDomainToHttpException() instead
-      throw new InternalServerErrorException(newFeatureOrError.left);
-    } else {
-      const result = await this.geoFeatureService.getById(
-        newFeatureOrError.right.id,
-      );
-      if (isNil(result)) {
-        // @debt Use mapDomainToHttpException() instead
-        throw new NotFoundException();
-      }
-
-      return this.geoFeatureSerializer.serialize(result);
-    }
-  }
-
-  @ImplementsAcl()
-  @ApiConsumesCsv({
-    description: 'Upload a csv with feature amount for each puid',
-  })
-  @ApiParam({
-    name: 'projectId',
-    description: 'Id of the Project the feature is part of',
-  })
-  @ApiOkResponse({ type: GeoFeature, isArray: true })
-  @UseInterceptors(
-    FileInterceptor('file', { limits: uploadOptions(50 * 1024 ** 2).limits }),
-  )
-  @Post(`:projectId/features/csv`)
-  async setFeatureAmountFromCSV(
-    @Param('projectId', ParseUUIDPipe) projectId: string,
-    @UploadedFile() file: Express.Multer.File,
-    @Req() req: RequestWithAuthenticatedUser,
-  ): Promise<GeoFeature[]> {
-    const result = await this.geoFeatureService.saveFeaturesFromCsv(
-      file.buffer,
-      projectId,
-      req.user.id,
-    );
-
-    if (isLeft(result)) {
-      throw mapAclDomainToHttpError(result.left, {
-        userId: req.user.id,
-        projectId,
-        resourceType: projectResource.name.plural,
-      });
-    }
-    return result.right;
-  }
-
-  @ImplementsAcl()
-  @ApiOperation({
-    description: 'Deletes a feature that is related to the given projectId',
-  })
-  @ApiParam({
-    name: 'projectId',
-    description: 'Id of the Project',
-  })
-  @ApiParam({
-    name: 'featureId',
-    description: 'ID of the Feature to be deleted',
-  })
-  @ApiTags(inlineJobTag)
-  @Delete(':projectId/features/:featureId')
-  async deleteFeature(
-    @Param('projectId') projectId: string,
-    @Param('featureId') featureId: string,
-    @Req() req: RequestWithAuthenticatedUser,
-  ): Promise<void> {
-    const result = await this.geoFeatureService.deleteFeature(
-      req.user.id,
-      projectId,
-      featureId,
-    );
-
-    if (isLeft(result)) {
-      switch (result.left) {
-        case projectNotFound:
-          throw new NotFoundException(`Project with id ${projectId} not found`);
-        case featureNotFound:
-          throw new NotFoundException(
-            `Feature with id ${featureId}, for project with id ${projectId}, not found`,
-          );
-        case featureIsLinkedToOneOrMoreScenarios:
-          throw new ForbiddenException(
-            `Feature with id ${featureId}, for project with id ${projectId}, still has Scenarios linked to it`,
-          );
-        case featureNotDeletable:
-          throw new ForbiddenException(
-            `Feature with id ${featureId}, for project with id ${projectId}, cannot be deleted`,
-          );
-      }
-    }
-
-    return;
-  }
-
-  @ImplementsAcl()
-  @ApiOperation({
-    description: `Updates A GeoFeature's tag for a given Project`,
-  })
-  @ApiParam({
-    name: 'projectId',
-    description: 'Id of the Project that the Feature is part of',
-  })
-  @ApiParam({
-    name: 'featureId',
-    description: 'Id of the Feature whose tag will be patched',
-  })
-  @ApiUnauthorizedResponse()
-  @ApiForbiddenResponse()
-  @ApiOkResponse({ type: GeoFeature })
-  @Patch(':projectId/features/:featureId/tags')
-  async updateGeoFeatureTag(
-    @Param('projectId', ParseUUIDPipe) projectId: string,
-    @Param('featureId', ParseUUIDPipe) featureId: string,
-    @Req() req: RequestWithAuthenticatedUser,
-    @Body() dto: UpdateGeoFeatureTagDTO,
-  ): Promise<GeoFeature> {
-    const result = await this.geoFeatureTagsService.setOrUpdateTagForFeature(
-      req.user.id,
-      projectId,
-      featureId,
-      dto.tagName,
-    );
-
-    if (isLeft(result)) {
-      throw mapAclDomainToHttpError(result.left, {
-        userId: req.user.id,
-        featureId,
-        projectId,
-      });
-    }
-
-    return this.geoFeatureService.serialize(result.right);
-  }
-
-  @ImplementsAcl()
-  @ApiOperation({
-    description: `Deletes A GeoFeature's tag for a given Project`,
-  })
-  @ApiParam({
-    name: 'projectId',
-    description: 'Id of the Project that the Feature is part of',
-  })
-  @ApiParam({
-    name: 'featureId',
-    description: 'Id of the Feature whose tag will be deleted',
-  })
-  @ApiUnauthorizedResponse()
-  @ApiForbiddenResponse()
-  @Delete(':projectId/features/:featureId/tags')
-  async deleteGeoFeatureTag(
-    @Param('featureId', ParseUUIDPipe) featureId: string,
-    @Param('projectId', ParseUUIDPipe) projectId: string,
-    @Req() req: RequestWithAuthenticatedUser,
-  ): Promise<void> {
-    const result = await this.geoFeatureTagsService.deleteTagForFeature(
-      req.user.id,
-      projectId,
-      featureId,
-    );
-
-    if (isLeft(result)) {
-      throw mapAclDomainToHttpError(result.left, {
-        userId: req.user.id,
-        featureId,
-        projectId,
-      });
-    }
-  }
-
-  @ImplementsAcl()
-  @ApiOperation({
-    description: `Returns all the tags for the given Project that match the filter`,
-  })
-  @ApiParam({
-    name: 'projectId',
-    description: 'Id of the Project that the tag to be removed pertains to',
-  })
-  @ApiUnauthorizedResponse()
-  @ApiForbiddenResponse()
-  @Get(':projectId/tags')
-  async getProjectTags(
-    @Param('projectId', ParseUUIDPipe) projectId: string,
-    @Req() req: RequestWithAuthenticatedUser,
-    @ProcessFetchSpecification() fetchSpecification: FetchSpecification,
-  ): Promise<GetProjectTagsResponseDto> {
-    const tagFilters = Array.isArray(fetchSpecification.filter?.tag)
-      ? fetchSpecification.filter?.tag
-      : undefined;
-    const result = await this.geoFeatureTagsService.getGeoFeatureTagsForProject(
-      req.user.id,
-      projectId,
-      tagFilters,
-      fetchSpecification.sort,
-    );
-
-    if (isLeft(result)) {
-      throw mapAclDomainToHttpError(result.left, {
-        userId: req.user.id,
-        projectId,
-      });
-    }
-
-    return { data: result.right };
-  }
-
-  @ImplementsAcl()
-  @ApiOperation({
-    description: `Updates the label of a tag for a given Project`,
-  })
-  @ApiParam({
-    name: 'projectId',
-    description: 'Id of the Project that the tag to be updated pertains to',
-  })
-  @ApiUnauthorizedResponse()
-  @ApiForbiddenResponse()
-  @Patch(':projectId/tags')
-  async updateProjectTag(
-    @Param('projectId', ParseUUIDPipe) projectId: string,
-    @Req() req: RequestWithAuthenticatedUser,
-    @Body() body: UpdateProjectTagDTO,
-  ): Promise<void> {
-    const result = await this.geoFeatureTagsService.updateTagForProject(
-      req.user.id,
-      projectId,
-      body,
-    );
-
-    if (isLeft(result)) {
-      throw mapAclDomainToHttpError(result.left, {
-        userId: req.user.id,
-        projectId,
-      });
-    }
-  }
-
-  @ImplementsAcl()
-  @ApiOperation({
-    description: `Deletes A Tag from a given Project, untagging all its features`,
-  })
-  @ApiParam({
-    name: 'projectId',
-    description: 'Id of the Project that the tag to be removed pertains to',
-  })
-  @ApiUnauthorizedResponse()
-  @ApiForbiddenResponse()
-  @Delete(':projectId/tags')
-  async deleteProjectTag(
-    @Param('projectId', ParseUUIDPipe) projectId: string,
-    @Req() req: RequestWithAuthenticatedUser,
-    @Body() dto: UpdateGeoFeatureTagDTO,
-  ): Promise<void> {
-    const result = await this.geoFeatureTagsService.deleteTagForProject(
-      req.user.id,
-      projectId,
-      dto.tagName,
-    );
-
-    if (isLeft(result)) {
-      throw mapAclDomainToHttpError(result.left, {
-        userId: req.user.id,
-        projectId,
-      });
-    }
   }
 
   @ImplementsAcl()
